+++ covquestions-editor-app/src/components/questionnaireEditor/QuestionnaireJsonEditor.tsx
@@ -10,8 +10,12 @@ import { JSONSchema7 } from "json-schema";
 import Ajv from "ajv";
 import { useAppDispatch } from "../../store/store";
 import { useSelector } from "react-redux";
-import { questionnaireJsonSelector, setHasErrors, setQuestionnaireInEditor } from "../../store/questionnaireInEditor";
 import { Grid } from "@material-ui/core";
+import {
+  questionnaireJsonSelector,
+  setHasErrorsInJsonMode,
+  setQuestionnaireInEditor,
+} from "../../store/questionnaireInEditor";
 
 type QuestionnaireFormEditorProps = {
   schema: JSONSchema7;
@@ -67,7 +71,7 @@ export const QuestionnaireJsonEditor: React.FC<QuestionnaireFormEditorProps> = (
         onChange={(newQuestionnaire: Questionnaire) => dispatch(setQuestionnaireInEditor(newQuestionnaire))}
         onValidationError={(errors: []) => {
           const hasErrors = errors.length > 0;
-          dispatch(setHasErrors(hasErrors));
+          dispatch(setHasErrorsInJsonMode(hasErrors));
         }}
       />
     </Grid>