+++ covquestions-editor-app/src/store/questionnaireInEditor.ts
@@ -20,7 +20,7 @@ type ArraySection =
   | SectionType.TEST_CASES;
 
 export const setQuestionnaireInEditor = createAction<EditorQuestionnaire>("setQuestionnaireInEditor");
-export const setHasErrors = createAction<boolean>("setHasErrors");
+export const setHasErrorsInJsonMode = createAction<boolean>("setHasErrorsInJsonMode");
 export const addNewQuestion = createAction("addNewQuestion");
 export const addNewResultCategory = createAction("addNewResultCategory");
 export const addNewVariable = createAction("addNewVariable");
@@ -64,7 +64,16 @@ export const editTestCase = createAction<{
 
 export type QuestionnaireWrapper = {
   questionnaire: EditorQuestionnaire;
-  hasErrors: boolean;
+  errors: {
+    isJsonModeError: boolean;
+    formEditor: {
+      meta: boolean;
+      questions: { [key: number]: boolean };
+      resultCategories: { [key: number]: boolean };
+      variables: { [key: number]: boolean };
+      testCases: { [key: number]: boolean };
+    };
+  };
 };
 
 const initialQuestionnaireInEditor: QuestionnaireWrapper = {
@@ -83,7 +92,16 @@ const initialQuestionnaireInEditor: QuestionnaireWrapper = {
     resultCategories: [],
     variables: [],
   },
-  hasErrors: false,
+  errors: {
+    isJsonModeError: false,
+    formEditor: {
+      meta: false,
+      questions: {},
+      resultCategories: {},
+      variables: {},
+      testCases: {},
+    },
+  },
 };
 
 export const questionnaireInEditor = createReducer(initialQuestionnaireInEditor, (builder) =>
@@ -91,26 +109,33 @@ export const questionnaireInEditor = createReducer(initialQuestionnaireInEditor,
     .addCase(setQuestionnaireInEditor, (state, action) => {
       state.questionnaire = addStringRepresentationToQuestionnaire(action.payload);
     })
-    .addCase(setHasErrors, (state, { payload }) => {
-      state.hasErrors = payload;
+    .addCase(setHasErrorsInJsonMode, (state, { payload }) => {
+      state.errors.isJsonModeError = payload;
+      state.errors.formEditor = {
+        meta: false,
+        questions: {},
+        resultCategories: {},
+        variables: {},
+        testCases: {},
+      };
     })
     .addCase(addNewQuestion, (state) => {
       state.questionnaire.questions.push({
-        id: "newQuestionId",
+        id: "new_question_id",
         text: "new question",
         type: "boolean",
       });
     })
     .addCase(addNewResultCategory, (state) => {
       state.questionnaire.resultCategories.push({
-        id: "newResultCategoryId",
+        id: "new_result_category_id",
         description: "",
         results: [],
       });
     })
     .addCase(addNewVariable, (state) => {
       state.questionnaire.variables.push({
-        id: "newVariableId",
+        id: "new_variable_id",
         expression: "",
       });
     })
@@ -128,6 +153,9 @@ export const questionnaireInEditor = createReducer(initialQuestionnaireInEditor,
       const questionnaireElement = state.questionnaire[section];
       if (questionnaireElement !== undefined) {
         questionnaireElement.splice(index, 1);
+        if (state.errors.formEditor[section][index] !== undefined) {
+          delete state.errors.formEditor[section][index];
+        }
       }
     })
     .addCase(swapItemWithNextOne, (state, { payload: { section, index } }) => {
@@ -136,11 +164,14 @@ export const questionnaireInEditor = createReducer(initialQuestionnaireInEditor,
         const tmp = questionnaireElement[index];
         questionnaireElement[index] = questionnaireElement[index + 1];
         questionnaireElement[index + 1] = tmp;
+        const tmpError = state.errors.formEditor[section][index] ?? false;
+        state.errors.formEditor[section][index] = state.errors.formEditor[section][index + 1] ?? false;
+        state.errors.formEditor[section][index + 1] = tmpError;
       }
     })
     .addCase(editMeta, (state, { payload: { changedMeta, hasErrors } }) => {
       state.questionnaire = moveRootPropertiesToQuestionnaire(state.questionnaire, changedMeta);
-      state.hasErrors = hasErrors;
+      state.errors.formEditor.meta = hasErrors;
     })
     .addCase(editQuestion, (state, { payload: { index, changedQuestion, hasErrors } }) => {
       let expression: LogicExpression | undefined = undefined;
@@ -151,7 +182,7 @@ export const questionnaireInEditor = createReducer(initialQuestionnaireInEditor,
         ...changedQuestion,
         enableWhenExpression: expression,
       };
-      state.hasErrors = hasErrors;
+      state.errors.formEditor.questions[index] = hasErrors;
     })
     .addCase(editResultCategory, (state, { payload: { index, changedResultCategory, hasErrors } }) => {
       state.questionnaire.resultCategories[index] = {
@@ -167,7 +198,7 @@ export const questionnaireInEditor = createReducer(initialQuestionnaireInEditor,
           };
         }),
       };
-      state.hasErrors = hasErrors;
+      state.errors.formEditor.resultCategories[index] = hasErrors;
     })
     .addCase(editVariable, (state, { payload: { index, changedVariable, hasErrors } }) => {
       let expression: LogicExpression = "";
@@ -178,14 +209,14 @@ export const questionnaireInEditor = createReducer(initialQuestionnaireInEditor,
         ...changedVariable,
         expression,
       };
-      state.hasErrors = hasErrors;
+      state.errors.formEditor.resultCategories[index] = hasErrors;
     })
     .addCase(editTestCase, (state, { payload: { index, changedTestCase, hasErrors } }) => {
       if (!state.questionnaire.testCases) {
         state.questionnaire.testCases = [];
       }
       state.questionnaire.testCases[index] = changedTestCase;
-      state.hasErrors = hasErrors;
+      state.errors.formEditor.testCases[index] = hasErrors;
     })
 );
 
@@ -204,4 +235,47 @@ export const variableInEditorSelector = (state: RootState, props: { index: numbe
 export const testCaseInEditorSelector = (state: RootState, props: { index: number }) => {
   const testCases = state.questionnaireInEditor.questionnaire.testCases;
   return testCases ? testCases[props.index] : undefined;
+};
+
+export const hasAnyErrorSelector = (state: RootState) => {
+  const questionnaireErrors = state.questionnaireInEditor.errors;
+  const currentErrorValues: boolean[] = [];
+  currentErrorValues.push(questionnaireErrors.isJsonModeError);
+  currentErrorValues.push(questionnaireErrors.formEditor.meta);
+
+  const sections = [
+    questionnaireErrors.formEditor.questions,
+    questionnaireErrors.formEditor.resultCategories,
+    questionnaireErrors.formEditor.variables,
+    questionnaireErrors.formEditor.testCases,
+  ];
+  for (const section of sections) {
+    for (const propertyName of Object.getOwnPropertyNames(section)) {
+      currentErrorValues.push(section[parseInt(propertyName)]);
+    }
+  }
+
+  return currentErrorValues.filter((it) => it).length > 0;
+};
+
+export const formErrorsSelector = (state: RootState) => state.questionnaireInEditor.errors.formEditor;
+
+export const duplicatedIdsSelector = (state: RootState) => {
+  const questionnaire = state.questionnaireInEditor.questionnaire;
+  const ids: string[] = [];
+  if (questionnaire.questions !== undefined) {
+    ids.push(...questionnaire.questions.map((it) => it.id));
+  }
+  if (questionnaire.resultCategories !== undefined) {
+    for (const resultCategory of questionnaire.resultCategories) {
+      ids.push(resultCategory.id);
+      if (resultCategory.results !== undefined) {
+        ids.push(...resultCategory.results.map((it) => it.id));
+      }
+    }
+  }
+  if (questionnaire.variables !== undefined) {
+    ids.push(...questionnaire.variables.map((it) => it.id));
+  }
+  return ids.filter((item, index) => item !== undefined && ids.indexOf(item) !== index);
 };