+++ covquestions-editor-app/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.7](https://github.com/CovOpen/CovQuestions/compare/covquestions-editor-app@1.0.6...covquestions-editor-app@1.0.7) (2021-03-09)
+
+**Note:** Version bump only for package covquestions-editor-app
+
+
+
+
+
+## [1.0.6](https://github.com/CovOpen/CovQuestions/compare/covquestions-editor-app@1.0.5...covquestions-editor-app@1.0.6) (2021-03-08)
+
+**Note:** Version bump only for package covquestions-editor-app
+
+
+
+
+
+## [1.0.5](https://github.com/CovOpen/CovQuestions/compare/covquestions-editor-app@1.0.4...covquestions-editor-app@1.0.5) (2021-03-08)
+
+**Note:** Version bump only for package covquestions-editor-app
+
+
+
+
+
 ## [1.0.4](https://github.com/CovOpen/CovQuestions/compare/covquestions-editor-app@1.0.3...covquestions-editor-app@1.0.4) (2021-03-07)
 
 **Note:** Version bump only for package covquestions-editor-app